refactor: rename utilities
